BitDiffer.Common/Detail/EnumDetail.cs: Include non-public fields of the enum class when adding EnumItemDetail
The fields of an enumeration class that represent the possible enum values are not necessarily public. In such a case, a NullReferenceException is thrown.

For instance the ccrewrite tool of Microsoft Code Contracts adds an enum (System.Diagnostics.Contracts.RuntimeContractsFlags) to every rewritten assembly that has internal enum fields.

This commit fixes this by including non-public fields in the type.GetField() call.