Fixes StackOverflowException for Assemblies that cannot be resolved
When running in Execution assembly load context a StackOverflowException
is thrown if an Assembly is not available to BitDiffer. This is because
it repeatedly invokes AppDomain.AssemblyResolve each time an
Assembly.Load is attempted. By unregistering the event handling
temporarily while we attempt Assembly.Load we can allow the failure to
occur without getting stuck in a loop.